+++ src/Symfony/Bundle/SecurityBundle/Security/FirewallMap.php
@@ -75,7 +75,7 @@ class FirewallMap implements FirewallMapInterface
                 /** @var FirewallContext $context */
                 $context = $this->container->get($contextId);
 
-                if ($context->getConfig()?->isStateless()) {
+                if ($context->getConfig()?->isStateless() && !$request->attributes->has('_stateless')) {
                     $request->attributes->set('_stateless', true);
                 }